adding error rate and restructuring tests
