+++ x-pack/test/apm_api_integration/basic/tests/transaction_groups/expectation/transaction_charts.json
